#8 Change folder names
