--- DeckOfCardsTests/Usings.cs
@@ -1,3 +1,0 @@
-global using NUnit.Framework;
-global using DeckOfPlayingCards;
-global using MathNet.Numerics.Distributions;
