+++ DeckOfPlayingCardsDemo/DeckOfPlayingCardsDemo.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\DeckOfPlayingCardsLibrary\DeckOfPlayingCardsLibrary.csproj" />
+  </ItemGroup>
+
+</Project>
