+++ DeckOfPlayingCardsTests/Usings.cs
@@ -1,0 +1,3 @@
+global using NUnit.Framework;
+global using DeckOfPlayingCards;
+global using MathNet.Numerics.Distributions;
